+++ tests/profiles/test_dlna.py
@@ -2,6 +2,7 @@
 
 from typing import List, Sequence
 
+from didl_lite import didl_lite
 import pytest
 
 from async_upnp_client import UpnpEventHandler, UpnpFactory, UpnpStateVariable
@@ -9,6 +10,7 @@ from async_upnp_client.client import UpnpService
 from async_upnp_client.profiles.dlna import (
     _parse_last_change_event,
     dlna_handle_notify_last_change,
+    DmrDevice,
 )
 
 from ..upnp_test_requester import RESPONSE_MAP, UpnpTestRequester
@@ -113,3 +115,285 @@ async def test_on_notify_dlna_event() -> None:
 
     state_var = service.state_variable("Volume")
     assert state_var.value == 50
+
+
+@pytest.mark.asyncio
+async def test_construct_play_media_metadata_types() -> None:
+    """Test various MIME and UPnP type options for construct_play_media_metadata."""
+    # pylint: disable=too-many-statements
+    requester = UpnpTestRequester(RESPONSE_MAP)
+    factory = UpnpFactory(requester)
+    device = await factory.async_create_device("http://dlna_dmr:1234/device.xml")
+    event_handler = UpnpEventHandler("http://localhost:11302", requester)
+    profile = DmrDevice(device, event_handler=event_handler)
+
+    media_url = "http://dlna_dms:4321/object/file_1222"
+    media_title = "Test music"
+
+    # No server to supply DLNA headers
+    metadata = didl_lite.from_xml_string(
+        await profile.construct_play_media_metadata(media_url, media_title)
+    )[0]
+    assert metadata.title == media_title
+    assert metadata.upnp_class == "object.item"
+    assert metadata.res[0].uri == media_url
+    assert metadata.res[0].protocol_info == "http-get:*:application/octet-stream:*"
+
+    metadata = didl_lite.from_xml_string(
+        await profile.construct_play_media_metadata(media_url + ".mp3", media_title)
+    )[0]
+    assert metadata.title == media_title
+    assert metadata.upnp_class == "object.item.audioItem"
+    assert metadata.res[0].uri == media_url + ".mp3"
+    assert metadata.res[0].protocol_info == "http-get:*:audio/mpeg:*"
+
+    metadata = didl_lite.from_xml_string(
+        await profile.construct_play_media_metadata(
+            media_url, media_title, default_mime_type="video/test-mime"
+        )
+    )[0]
+    assert metadata.title == media_title
+    assert metadata.upnp_class == "object.item.videoItem"
+    assert metadata.res[0].uri == media_url
+    assert metadata.res[0].protocol_info == "http-get:*:video/test-mime:*"
+
+    metadata = didl_lite.from_xml_string(
+        await profile.construct_play_media_metadata(
+            media_url, media_title, default_upnp_class="object.item.imageItem"
+        )
+    )[0]
+    assert metadata.title == media_title
+    assert metadata.upnp_class == "object.item.imageItem"
+    assert metadata.res[0].uri == media_url
+    assert metadata.res[0].protocol_info == "http-get:*:application/octet-stream:*"
+
+    metadata = didl_lite.from_xml_string(
+        await profile.construct_play_media_metadata(
+            media_url, media_title, override_mime_type="video/test-mime"
+        )
+    )[0]
+    assert metadata.title == media_title
+    assert metadata.upnp_class == "object.item.videoItem"
+    assert metadata.res[0].uri == media_url
+    assert metadata.res[0].protocol_info == "http-get:*:video/test-mime:*"
+
+    metadata = didl_lite.from_xml_string(
+        await profile.construct_play_media_metadata(
+            media_url, media_title, override_upnp_class="object.item.imageItem"
+        )
+    )[0]
+    assert metadata.title == media_title
+    assert metadata.upnp_class == "object.item.imageItem"
+    assert metadata.res[0].uri == media_url
+    assert metadata.res[0].protocol_info == "http-get:*:application/octet-stream:*"
+
+    metadata = didl_lite.from_xml_string(
+        await profile.construct_play_media_metadata(
+            media_url,
+            media_title,
+            override_dlna_features="DLNA_OVERRIDE_FEATURES",
+        )
+    )[0]
+    assert metadata.title == media_title
+    assert metadata.upnp_class == "object.item"
+    assert metadata.res[0].uri == media_url
+    assert (
+        metadata.res[0].protocol_info
+        == "http-get:*:application/octet-stream:DLNA_OVERRIDE_FEATURES"
+    )
+
+    metadata = didl_lite.from_xml_string(
+        await profile.construct_play_media_metadata(
+            media_url,
+            media_title,
+            override_mime_type="video/test-mime",
+            override_dlna_features="DLNA_OVERRIDE_FEATURES",
+        )
+    )[0]
+    assert metadata.title == media_title
+    assert metadata.upnp_class == "object.item.videoItem"
+    assert metadata.res[0].uri == media_url
+    assert (
+        metadata.res[0].protocol_info
+        == "http-get:*:video/test-mime:DLNA_OVERRIDE_FEATURES"
+    )
+
+    # Media server supplies media information for HEAD requests
+    requester.response_map[("HEAD", media_url)] = (
+        200,
+        {
+            "ContentFeatures.dlna.org": "DLNA_SERVER_FEATURES",
+            "Content-Type": "video/server-mime",
+        },
+        "",
+    )
+    requester.response_map[("HEAD", media_url + ".mp3")] = (
+        200,
+        {
+            "ContentFeatures.dlna.org": "DLNA_SERVER_FEATURES",
+            "Content-Type": "video/server-mime",
+        },
+        "",
+    )
+
+    metadata = didl_lite.from_xml_string(
+        await profile.construct_play_media_metadata(media_url, media_title)
+    )[0]
+    assert metadata.title == media_title
+    assert metadata.upnp_class == "object.item.videoItem"
+    assert metadata.res[0].uri == media_url
+    assert (
+        metadata.res[0].protocol_info
+        == "http-get:*:video/server-mime:DLNA_SERVER_FEATURES"
+    )
+
+    metadata = didl_lite.from_xml_string(
+        await profile.construct_play_media_metadata(media_url + ".mp3", media_title)
+    )[0]
+    assert metadata.title == media_title
+    assert metadata.upnp_class == "object.item.videoItem"
+    assert metadata.res[0].uri == media_url + ".mp3"
+    assert (
+        metadata.res[0].protocol_info
+        == "http-get:*:video/server-mime:DLNA_SERVER_FEATURES"
+    )
+
+    metadata = didl_lite.from_xml_string(
+        await profile.construct_play_media_metadata(
+            media_url, media_title, default_mime_type="video/test-mime"
+        )
+    )[0]
+    assert metadata.title == media_title
+    assert metadata.upnp_class == "object.item.videoItem"
+    assert metadata.res[0].uri == media_url
+    assert (
+        metadata.res[0].protocol_info
+        == "http-get:*:video/server-mime:DLNA_SERVER_FEATURES"
+    )
+
+    metadata = didl_lite.from_xml_string(
+        await profile.construct_play_media_metadata(
+            media_url, media_title, default_upnp_class="object.item.imageItem"
+        )
+    )[0]
+    assert metadata.title == media_title
+    assert metadata.upnp_class == "object.item.videoItem"
+    assert metadata.res[0].uri == media_url
+    assert (
+        metadata.res[0].protocol_info
+        == "http-get:*:video/server-mime:DLNA_SERVER_FEATURES"
+    )
+
+    metadata = didl_lite.from_xml_string(
+        await profile.construct_play_media_metadata(
+            media_url, media_title, override_mime_type="image/test-mime"
+        )
+    )[0]
+    assert metadata.title == media_title
+    assert metadata.upnp_class == "object.item.imageItem"
+    assert metadata.res[0].uri == media_url
+    assert (
+        metadata.res[0].protocol_info
+        == "http-get:*:image/test-mime:DLNA_SERVER_FEATURES"
+    )
+
+    metadata = didl_lite.from_xml_string(
+        await profile.construct_play_media_metadata(
+            media_url, media_title, override_upnp_class="object.item.imageItem"
+        )
+    )[0]
+    assert metadata.title == media_title
+    assert metadata.upnp_class == "object.item.imageItem"
+    assert metadata.res[0].uri == media_url
+    assert (
+        metadata.res[0].protocol_info
+        == "http-get:*:video/server-mime:DLNA_SERVER_FEATURES"
+    )
+
+    metadata = didl_lite.from_xml_string(
+        await profile.construct_play_media_metadata(
+            media_url,
+            media_title,
+            override_dlna_features="DLNA_OVERRIDE_FEATURES",
+        )
+    )[0]
+    assert metadata.title == media_title
+    assert metadata.upnp_class == "object.item.videoItem"
+    assert metadata.res[0].uri == media_url
+    assert (
+        metadata.res[0].protocol_info
+        == "http-get:*:video/server-mime:DLNA_OVERRIDE_FEATURES"
+    )
+
+    metadata = didl_lite.from_xml_string(
+        await profile.construct_play_media_metadata(
+            media_url,
+            media_title,
+            override_mime_type="image/test-mime",
+            override_dlna_features="DLNA_OVERRIDE_FEATURES",
+        )
+    )[0]
+    assert metadata.title == media_title
+    assert metadata.upnp_class == "object.item.imageItem"
+    assert metadata.res[0].uri == media_url
+    assert (
+        metadata.res[0].protocol_info
+        == "http-get:*:image/test-mime:DLNA_OVERRIDE_FEATURES"
+    )
+
+
+@pytest.mark.asyncio
+async def test_construct_play_media_metadata_meta_data() -> None:
+    """Test meta_data values for construct_play_media_metadata."""
+    requester = UpnpTestRequester(RESPONSE_MAP)
+    factory = UpnpFactory(requester)
+    device = await factory.async_create_device("http://dlna_dmr:1234/device.xml")
+    event_handler = UpnpEventHandler("http://localhost:11302", requester)
+    profile = DmrDevice(device, event_handler=event_handler)
+
+    media_url = "http://dlna_dms:4321/object/file_1222.mp3"
+    media_title = "Test music"
+    meta_data = {
+        "title": "Test override title",  # Should override media_title parameter
+        "description": "Short test description",  # In base audioItem class
+        "artist": "Test singer",
+        "album": "Test album",
+        "originalTrackNumber": 3,  # Should be converted to lower_camel_case
+    }
+
+    # No server information about media type or contents
+
+    # Without specifying UPnP class, only generic types lacking certain values are used
+    metadata = didl_lite.from_xml_string(
+        await profile.construct_play_media_metadata(
+            media_url,
+            media_title,
+            meta_data=meta_data,
+        )
+    )[0]
+    assert metadata.upnp_class == "object.item.audioItem"
+    assert metadata.title == "Test override title"
+    assert metadata.description == "Short test description"
+    assert not hasattr(metadata, "artist")
+    assert not hasattr(metadata, "album")
+    assert not hasattr(metadata, "original_track_number")
+    assert metadata.res[0].uri == media_url
+    assert metadata.res[0].protocol_info == "http-get:*:audio/mpeg:*"
+
+    # Set the UPnP class correctly
+    metadata = didl_lite.from_xml_string(
+        await profile.construct_play_media_metadata(
+            media_url,
+            media_title,
+            override_upnp_class="object.item.audioItem.musicTrack",
+            meta_data=meta_data,
+        )
+    )[0]
+    assert metadata.upnp_class == "object.item.audioItem.musicTrack"
+    assert metadata.title == "Test override title"
+    assert metadata.description == "Short test description"
+    assert metadata.artist == "Test singer"
+    assert metadata.album == "Test album"
+    assert metadata.original_track_number == "3"
+    assert metadata.res[0].uri == media_url
+    assert metadata.res[0].protocol_info == "http-get:*:audio/mpeg:*"